rename generate files
