+++ typescript/schemas/src/internal/generateSdkCpp.ts
