+++ typescript/schemas/src/internal/index.ts
@@ -3,6 +3,5 @@ export * from "./generateTypeScript";
 export * from "./generateJsonSchema";
 export * from "./generateProto";
 export * from "./generateRos";
-export * from "./generateRust";
 export * from "./schemas";
 export * from "./types";
